requirements.txt
discord.py==2.5.2
python-dotenv==0.20.0
requests==2.32.3
SpeechRecognition==3.14.2
pydub==0.25.1
ffmpeg-python==0.2.0
openai-whisper==20240930